refactor: step object로 관리
